Ft#115546793 pushup sensor

+++ app/src/main/java/com/example/andela/fitgoup/model/PushUpModel.java
@@ -19,7 +19,7 @@ import java.util.List;
 public class PushUpModel extends Model {
 
   @Column(name = "day_pushup")
-  public int pushups;
+  public long pushups;
 
   public PushUpModel () {
     super();
